+++ src/main/java/org/co2/kanban/rest/project/task/TaskAssembler.java
@@ -34,8 +34,8 @@ public class TaskAssembler extends ResourceAssemblerSupport<Task, TaskResource> 
         super(TaskController.class, TaskResource.class);
     }
 
-    private Float calculateTimeRemains(Float currentTimeRemains, Allocation allocation) {
-        Float timeRemains = currentTimeRemains;
+    private Double calculateTimeRemains(Double currentTimeRemains, Allocation allocation) {
+        Double timeRemains = currentTimeRemains;
         if (allocation.getTimeRemains() != null) {
             timeRemains = allocation.getTimeRemains();
         } else {
@@ -47,8 +47,8 @@ public class TaskAssembler extends ResourceAssemblerSupport<Task, TaskResource> 
     @Override
     public TaskResource toResource(Task task) {
         TaskResource resource = new TaskResource(task);
-        Float timeSpent = 0F;
-        Float timeRemains = 0F;
+        Double timeSpent = 0D;
+        Double timeRemains = 0D;
         if(task.getEstimatedLoad() != null) {
             timeRemains = task.getEstimatedLoad();
         }
